+++ packages/fake-data/src/populateDb/helpers/encounter.ts
